scripts/requirements.txt
pandas>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
boto3>=1.35
requests>=2.31
